creating saved object for association
